backend/requirements.txt
django-oauth-toolkit==2.3.0
jwcrypto==1.5.0
wrapt==1.15.0
pillow==9.3.0
djangorestframework-simplejwt==5.2.2